기간결산 일부부 완료 및 기초관리 수정

+++ src/main/java/egovframework/veterans/com/cmm/VtcSLOrderController.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -13,13 +14,20 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.SessionAttributes;
 
 import egovframework.veterans.com.cmm.service.VtcSLOrderService;
 import egovframework.veterans.com.cmm.service.vo.SLOrderGroup;
 import egovframework.veterans.com.cmm.service.vo.SLOrderItem;
+import egovframework.veterans.com.cmm.service.vo.Users;
+import lombok.RequiredArgsConstructor;
 
 @Controller
+@RequiredArgsConstructor
+@SessionAttributes("loginuserinfo")
 public class VtcSLOrderController {
+	
+	private final HttpSession session;
 
 	private ApplicationContext applicationContext;
 	
@@ -33,14 +41,21 @@ public class VtcSLOrderController {
 		LOGGER.info("VtcPaidController setApplicationContext method has called!");
 	}
 	
+	
+	
 	@Resource(name="VtcSLOrderService")
 	protected VtcSLOrderService VtcSLService;
 	
 	
 	@RequestMapping(value="SLOrderGroup.do")
-	public String selectSLOrder(String SiteCode, ModelMap model) throws Exception {
-		SiteCode = "10001";
-		List<SLOrderGroup> list = VtcSLService.selectSLOrderGroup(SiteCode);
+	public String selectSLOrder(ModelMap model) throws Exception {
+		Users users = (Users) session.getAttribute("loginuserinfo");
+		if(users == null){
+			model.addAttribute("msg", "로그인을 다시 해주세요.");
+			model.addAttribute("script", "back");
+			return "redirect:login.do";
+		}
+		List<SLOrderGroup> list = VtcSLService.selectSLOrderGroup(users.getSiteCode());
 		model.addAttribute("list", list);
 		return "basic/orders/orderGroup";
 	}
@@ -49,60 +64,63 @@ public class VtcSLOrderController {
 	public String orderGroupModify(SLOrderGroup group, ModelMap model) throws Exception {
 		group = VtcSLService.getOrderGroupDetail(group);
 		
-		System.out.println("group : " + group);
 		
 		model.addAttribute("group", group);
 		return "basic/orders/orderGroup_modity";
 	}
 	@RequestMapping(value="OrderGroupUpdOK.do")
-	public String orderGroupModifyOK(HttpServletRequest request, SLOrderGroup group) throws Exception {
-		String SiteCode = request.getParameter("SiteCode");
-		int pkid = Integer.parseInt(request.getParameter("pkid"));
-		String GroupName = request.getParameter("GroupName");
-		int GroupJungWon = Integer.parseInt(request.getParameter("GroupJungWon"));
-		int SortOrder = Integer.parseInt(request.getParameter("SortOrder"));
-		String ParkingTimeWeek = request.getParameter("ParkingTimeWeek");
-		String ParkingTimeWeekend = request.getParameter("ParkingTimeWeekend");
-		String UpdDate = request.getParameter("UpdDate");
+	public String orderGroupModifyOK(ModelMap model, SLOrderGroup group) throws Exception {
+		Users users = (Users) session.getAttribute("loginuserinfo");
+		if(users == null){
+			model.addAttribute("msg", "로그인을 다시 해주세요.");
+			model.addAttribute("script", "back");
+			return "redirect:login.do";
+		}
 		
-		group = new SLOrderGroup();
-		group.setSiteCode(SiteCode);
-		group.setPkid(pkid);
-		group.setGroupName(GroupName);
-		group.setGroupJungWon(GroupJungWon);
-		group.setSortOrder(SortOrder);
-		group.setParkingTimeWeek(ParkingTimeWeek);
-		group.setParkingTimeWeekend(ParkingTimeWeekend);
-		group.setUpdDate(UpdDate);
+		group.setSiteCode(users.getSiteCode());
+		group.setUpdDatePKID(users.getUserPKID());
 		VtcSLService.updateOrderGroup(group);
 		return "redirect:SLOrderGroup.do";
 	}
 	
 	@RequestMapping(value="OrderGpInsert.do")
-	public String orderGroupInsert(HttpServletRequest request, SLOrderGroup group) throws Exception {
+	public String orderGroupInsert(ModelMap model) throws Exception {
+		
+		Users users = (Users) session.getAttribute("loginuserinfo");
+		if(users == null){
+			model.addAttribute("msg", "로그인을 다시 해주세요.");
+			model.addAttribute("script", "back");
+			return "redirect:login.do";
+		}
+		int SortOrder = VtcSLService.getGroupSortOrder(users.getSiteCode());
+		model.addAttribute("sortorder", SortOrder + 1);
 		return "basic/orders/orderGroup_insert";
 	}
+	
 	@RequestMapping(value="OrderGpInsertOK.do")
-	public String orderGroupInsertOK(HttpServletRequest request, SLOrderGroup group) throws Exception {
-		String SiteCode = "10001";
-		String GroupName = request.getParameter("GroupName");
-		int GroupJungWon = Integer.parseInt(request.getParameter("GroupJungWon"));
-		String ParkingTimeWeek = request.getParameter("ParkingTimeWeek");
-		String ParkingTimeWeekend = request.getParameter("ParkingTimeWeekend");
-		int SortOrder = Integer.parseInt(request.getParameter("SortOrder"));
+	public String orderGroupInsertOK(SLOrderGroup group, ModelMap model) throws Exception {
+		Users users = (Users) session.getAttribute("loginuserinfo");
+		if(users == null){
+			model.addAttribute("msg", "로그인을 다시 해주세요.");
+			model.addAttribute("script", "back");
+			return "redirect:login.do";
+		}
+		group.setSiteCode(users.getSiteCode());
+		group.setAddUserPKID(users.getUserPKID());
 		
-		group = new SLOrderGroup();
-		group.setSiteCode(SiteCode);
-		group.setGroupName(GroupName);
-		group.setGroupJungWon(GroupJungWon);
-		group.setParkingTimeWeek(ParkingTimeWeek);
-		group.setParkingTimeWeekend(ParkingTimeWeekend);
-		group.setSortOrder(SortOrder);
 		VtcSLService.insertOrderGroup(group);
 		return "redirect:SLOrderGroup.do";
 	}
 	@RequestMapping(value="deleteSlGp.do")
 	public String deleteOrderGroup(ModelMap model, SLOrderGroup group) throws Exception {
+		Users users = (Users) session.getAttribute("loginuserinfo");
+		if(users == null){
+			model.addAttribute("msg", "로그인을 다시 해주세요.");
+			model.addAttribute("script", "back");
+			return "redirect:login.do";
+		}
+		group.setUpdDatePKID(users.getUserPKID());
+		
 		VtcSLService.deleteOrderGroup(group);
 		return "redirect:SLOrderGroup.do";
 	}
@@ -110,14 +128,21 @@ public class VtcSLOrderController {
 	
 	
 	@RequestMapping(value="SLOrderItem.do")
-	public String selectSLOrderItem(String SiteCode, ModelMap model) throws Exception {
-		SiteCode = "10001";
-		List<SLOrderItem> list = VtcSLService.listSLOderItem(SiteCode);
+	public String selectSLOrderItem(ModelMap model) throws Exception {
+		Users users = (Users) session.getAttribute("loginuserinfo");
+		if(users == null){
+			model.addAttribute("msg", "로그인을 다시 해주세요.");
+			model.addAttribute("script", "back");
+			return "redirect:login.do";
+		}
+		
+		List<SLOrderItem> list = VtcSLService.listSLOderItem(users.getSiteCode());
 		model.addAttribute("list", list);
 		return "basic/orders/orderItem";
 	}
 	@RequestMapping(value="OrderItemUpd.do")
 	public String orderItemModify(SLOrderItem item, ModelMap model, HttpServletRequest request) throws Exception {
+		
 		String SiteCode = request.getParameter("SiteCode");
 		List<SLOrderGroup> list = VtcSLService.selectSLOrderGroup(SiteCode);
 		
@@ -127,125 +152,54 @@ public class VtcSLOrderController {
 		model.addAttribute("item", item);
 		return "basic/orders/orderItem_modity";
 	}
+	
 	@RequestMapping(value="OrderItemUpdOK.do")
-	public String orderItemModifyOK(HttpServletRequest request, SLOrderItem item) throws Exception {
-		String SiteCode = request.getParameter("SiteCode");
-		int pkid = Integer.parseInt(request.getParameter("pkid"));
-		String ItemName = request.getParameter("ItemName");
-		String ShortName = request.getParameter("ShortName");
-		int GroupCode = Integer.parseInt(request.getParameter("GroupCode"));
-		String FromTime = request.getParameter("FromTime");
-		String ToTime = request.getParameter("ToTime");
-		String AdultGBN = request.getParameter("AdultGBN");
-		String Gender = request.getParameter("Gender");
-		String dcType = request.getParameter("dcType");
-		String Price = request.getParameter("Price");
-		String vat = request.getParameter("vat");
-		String Nvat = request.getParameter("Nvat");
-		int Jungwon = Integer.parseInt(request.getParameter("Jungwon"));
-		String Mon = request.getParameter("Mon");
-		String Tue = request.getParameter("Tue");
-		String Wed = request.getParameter("Wed");
-		String Thu = request.getParameter("Thu");
-		String Fri = request.getParameter("Fri");
-		String Sat = request.getParameter("Sat");
-		String Sun = request.getParameter("Sun");
-		String UpJang = request.getParameter("UpJang");
-		String KioskYN = request.getParameter("KioskYN");
-		int SortOrder = Integer.parseInt(request.getParameter("SortOrder"));
-		String UpdDate = request.getParameter("UpdDate");
+	public String orderItemModifyOK( SLOrderItem item, ModelMap model) throws Exception {
 		
-		item = new SLOrderItem();
-		item.setSiteCode(SiteCode);
-		item.setPkid(pkid);
-		item.setItemName(ItemName);
-		item.setShortName(ShortName);
-		item.setGroupCode(GroupCode);
-		item.setFromTime(FromTime);
-		item.setToTime(ToTime);
-		item.setAdultGBN(AdultGBN);
-		item.setGender(Gender);
-		item.setDcType(dcType);
-		item.setPrice(Price);
-		item.setVat(vat);
-		item.setNvat(Nvat);
-		item.setJungwon(Jungwon);
-		item.setMon(Mon);
-		item.setTue(Tue);
-		item.setWed(Wed);
-		item.setThu(Thu);
-		item.setFri(Fri);
-		item.setSat(Sat);
-		item.setSun(Sun);
-		item.setUpJang(UpJang);
-		item.setKioskYN(KioskYN);
-		item.setSortOrder(SortOrder);
-		item.setUpdDate(UpdDate);
+		Users users = (Users) session.getAttribute("loginuserinfo");
+		if(users == null){
+			model.addAttribute("msg", "로그인을 다시 해주세요.");
+			model.addAttribute("script", "back");
+			return "redirect:login.do";
+		}
+		item.setSiteCode(users.getSiteCode());
+		item.setUpdUserPKID(users.getUserPKID());
+		
 		VtcSLService.updateOrderItem(item);
 		
 		return "redirect:SLOrderItem.do";
 	}
 	
 	@RequestMapping(value="OrderItemInsert.do")
-	public String OrderItemInsert(HttpServletRequest request, ModelMap model, String SiteCode) throws Exception {
-		SiteCode = "10001";
-		List<SLOrderGroup> list = VtcSLService.selectSLOrderGroup(SiteCode);
+	public String OrderItemInsert(ModelMap model) throws Exception {
+		Users users = (Users) session.getAttribute("loginuserinfo");
+		if(users == null){
+			model.addAttribute("msg", "로그인을 다시 해주세요.");
+			model.addAttribute("script", "back");
+			return "redirect:login.do";
+		}
+		
+		int SortOrder = VtcSLService.getItemSortOrder(users.getSiteCode());
+		
+		List<SLOrderGroup> list = VtcSLService.selectSLOrderGroup(users.getSiteCode());
+		
+		model.addAttribute("sortorder", SortOrder + 1);
 		model.addAttribute("list", list);
 		return "basic/orders/orderItem_insert";
 	}
 	
 	@RequestMapping(value="OrderItemInsertOk.do")
-	public String OrderItemInsertOK(HttpServletRequest request, SLOrderItem item) throws Exception {
-		String SiteCode = "10001";
-		String ItemName = request.getParameter("ItemName");
-		String ShortName = request.getParameter("ShortName");
-		int GroupCode = Integer.parseInt(request.getParameter("GroupCode"));
-		String FromTime = request.getParameter("FromTime");
-		String ToTime = request.getParameter("ToTime");
-		String AdultGBN = request.getParameter("AdultGBN");
-		String Gender = request.getParameter("Gender");
-		String dcType = request.getParameter("dcType");
-		String Price = request.getParameter("Price");
-		String vat = request.getParameter("vat");
-		String Nvat = request.getParameter("Nvat");
-		int Jungwon = Integer.parseInt(request.getParameter("Jungwon"));
-		String Mon = request.getParameter("Mon");
-		String Tue = request.getParameter("Tue");
-		String Wed = request.getParameter("Wed");
-		String Thu = request.getParameter("Thu");
-		String Fri = request.getParameter("Fri");
-		String Sat = request.getParameter("Sat");
-		String Sun = request.getParameter("Sun");
-		String UpJang = request.getParameter("UpJang");
-		String KioskYN = request.getParameter("KioskYN");
-		int SortOrder = Integer.parseInt(request.getParameter("SortOrder"));
-		String UpdDate = request.getParameter("UpdDate");
+	public String OrderItemInsertOK(ModelMap model, SLOrderItem item) throws Exception {
+		Users users = (Users) session.getAttribute("loginuserinfo");
+		if(users == null){
+			model.addAttribute("msg", "로그인을 다시 해주세요.");
+			model.addAttribute("script", "back");
+			return "redirect:login.do";
+		}
+
+		item.setSiteCode(users.getSiteCode());
+		item.setAddUserPKID(users.getUserPKID());
 		
-		item = new SLOrderItem();
-		item.setSiteCode(SiteCode);
-		item.setItemName(ItemName);
-		item.setShortName(ShortName);
-		item.setGroupCode(GroupCode);
-		item.setFromTime(FromTime);
-		item.setToTime(ToTime);
-		item.setAdultGBN(AdultGBN);
-		item.setGender(Gender);
-		item.setDcType(dcType);
-		item.setPrice(Price);
-		item.setVat(vat);
-		item.setNvat(Nvat);
-		item.setJungwon(Jungwon);
-		item.setMon(Mon);
-		item.setTue(Tue);
-		item.setWed(Wed);
-		item.setThu(Thu);
-		item.setFri(Fri);
-		item.setSat(Sat);
-		item.setSun(Sun);
-		item.setUpJang(UpJang);
-		item.setKioskYN(KioskYN);
-		item.setSortOrder(SortOrder);
-		item.setUpdDate(UpdDate);
 		VtcSLService.insertOrderItem(item);
 		return "redirect:SLOrderItem.do";
 	}
